requirements.txt
django
gunicorn
psycopg2-binary
WhiteNoise
python-decouple
cryptography
djangorestframework
djangorestframework-simplejwt
markdown
django-filter
drf-yasg
django-cors-headers
drf_spectacular
